requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
PyMySQL==1.1.0
boto3==1.34.0
Pillow==10.1.0
Flask-CORS==4.0.0
cryptography>=41.0.0 